# dashboard/.schema-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Orchestrator Dashboard",
        "description": "REST/SSE API of the YAOE-FLOW observability dashboard.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "operationId": "getApiNotifications",
                "tags": [
                    "Notifications"
                ],
                "summary": "List channels with rules and masked config",
                "responses": {
                    "200": {
                        "description": "Canais",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "postApiNotifications",
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a notification channel",
                "responses": {
                    "200": {
                        "description": "Canal criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "channel": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "channel"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de valida\u00e7\u00e3o",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "config": {
                                        "type": "object",
                                        "propertyNames": {
                                            "type": "string"
                                        },
                                        "additionalProperties": {}
                                    }
                                },
                                "required": [
                                    "type",
                                    "name"
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "patch": {
                "operationId": "patchApiNotificationsById",
                "tags": [
                    "Notifications"
                ],
                "summary": "Update a channel",
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "channel": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "channel"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "config": {
                                        "type": "object",
                                        "propertyNames": {
                                            "type": "string"
                                        },
                                        "additionalProperties": {}
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteApiNotificationsById",
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete a channel",
                "responses": {
                    "200": {
                        "description": "Removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ]
            }
        },
        "/api/notifications/{id}/rules/{event}": {
            "put": {
                "operationId": "putApiNotificationsByIdRulesByEvent",
                "tags": [
                    "Notifications"
                ],
                "summary": "Toggle an event on a channel",
                "responses": {
                    "200": {
                        "description": "Regra atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "enabled": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "event",
                        "required": true
                    }
                ]
            }
        },
        "/api/notifications/{id}/test": {
            "post": {
                "operationId": "postApiNotificationsByIdTest",
                "tags": [
                    "Notifications"
                ],
                "summary": "Test a channel with a real message",
                "responses": {
                    "200": {
                        "description": "Resultado do teste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Canal n\u00e3o encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ]
            }
        },
...
    },
...
}